+++ src/sgr/color/types.d.ts
@@ -1,0 +1,28 @@
+export interface Channel {
+	black   : SGR;
+	red     : SGR;
+	green   : SGR;
+	yellow  : SGR;
+	blue    : SGR;
+	magenta : SGR;
+	cyan    : SGR;
+	white   : SGR;
+
+	brightBlack   : SGR;
+	brightRed     : SGR;
+	brightGreen   : SGR;
+	brightYellow  : SGR;
+	brightBlue    : SGR;
+	brightMagenta : SGR;
+	brightCyan    : SGR;
+	brightWhite   : SGR;
+
+	rgb: (r: number, g: number, b: number) => SGR;
+
+	x256: {
+		(code: number): SGR;
+		(r: number, g: number, b: number): SGR;
+	};
+
+	default: string,
+}